Stop using std::mutex on Windows builds.
Reworked the threading code to an oslib and fixed some use of it.
Had to add windows.h includes and some other minor refactor since
some file were relaying on that header being already included.
While we are at it replace min/max functions with std:: ones.
Seems that we were using the windef.h defined macros (and no idea
where they come from in Linux, which sounds scary to me).
