+++ CMakeLists.txt
@@ -179,6 +179,7 @@ if (${HOST_OS} EQUAL ${OS_WINDOWS})
 
   list(APPEND osd_SRCS ${d_core}/windows/winmain.cpp)
   list(APPEND osd_SRCS ${d_core}/windows/win_vmem.cpp)
+  list(APPEND osd_SRCS ${d_core}/oslib/windows/threading.cpp)
   list(APPEND osd_SRCS ${d_aout}/audiobackend_directsound.cpp)
   
   
@@ -190,6 +191,7 @@ elseif (${HOST_OS} EQUAL ${OS_LINUX} OR ${HOST_OS} EQUAL ${OS_ANDROID})
     list(APPEND osd_SRCS 
       ${d_core}/linux/common.cpp
       ${d_core}/linux/context.cpp
+      ${d_core}/linux/oslib/posix/threading.cpp
       ${d_core}/linux/nixprof/nixprof.cpp
 
       ${d_aout}/audiobackend_oss.cpp # add option
