+++ core/hw/naomi/gdcartridge.cpp
@@ -594,7 +594,7 @@ void GDCartridge::device_reset()
 void *GDCartridge::GetDmaPtr(u32 &size)
 {
 	dimm_cur_address = DmaOffset & (dimm_data_size-1);
-	size = min(size, dimm_data_size - dimm_cur_address);
+	size = std::min(size, dimm_data_size - dimm_cur_address);
 	return dimm_data + dimm_cur_address;
 }
 
@@ -608,7 +608,7 @@ void GDCartridge::AdvancePtr(u32 size)
 bool GDCartridge::Read(u32 offset, u32 size, void *dst)
 {
 	u32 addr = offset & (dimm_data_size-1);
-	memcpy(dst, &dimm_data[addr], min(size, dimm_data_size - addr));
+	memcpy(dst, &dimm_data[addr], std::min(size, dimm_data_size - addr));
 	return true;
 }
