+++ core/hw/sh4/dyna/decoder.cpp
@@ -1185,9 +1185,9 @@ _end:
 	}
 
 	//make sure we don't use wayy-too-many cycles
-	blk->guest_cycles=min(blk->guest_cycles,max_cycles);
+	blk->guest_cycles=std::min(blk->guest_cycles,max_cycles);
 	//make sure we don't use wayy-too-few cycles
-	blk->guest_cycles=max(1U,blk->guest_cycles);
+	blk->guest_cycles=std::max(1U,blk->guest_cycles);
 	blk=0;
 }
