+++ core/windows/win_keyboard.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "input/keyboard_device.h"
+#include <windows.h>
 
 // Used to differentiate between main enter key and num keypad one
 #define VK_NUMPAD_RETURN 0x0E
