+++ core/windows/winmain.cpp
@@ -1,3 +1,8 @@
+
+#define _WIN32_WINNT 0x0502
+#include <windows.h>
+#include <windowsx.h>
+
 #include "oslib/oslib.h"
 #include "oslib/audiostream.h"
 #include "imgread/common.h"
@@ -5,10 +10,6 @@
 #include "cfg/cfg.h"
 #include "xinput_gamepad.h"
 #include "win_keyboard.h"
-
-#define _WIN32_WINNT 0x0500
-#include <windows.h>
-#include <windowsx.h>
 
 #include <xinput.h>
 #include "hw/maple/maple_cfg.h"
